fix: address PR review findings (round 1)
- list_traject_document_jobs: LIMIT 100 so a traject with many failed uploads
  can't grow the status query/UI without bound
- extract the shared hidden-file-input upload wiring into useDocumentUpload,
  used by both the launcher sheet and the standalone page (was duplicated)
